Relocating contianer pages
