+++ components/blocks/accordion-block/index.tsx
@@ -1,13 +1,8 @@
-"use client";
+import type { FC } from "react";
 
-import { useTranslations } from "next-intl";
-import { type FC, useContext, useState } from "react";
-
-import CrossIcon from "@/assets/icons/cross.svg";
-import PlusIcon from "@/assets/icons/plus.svg";
 import { HtmlParser } from "@/components/typography/html-parser";
 import type { FaqFragment as IFaq } from "@/graphql/__generated__/operations";
-import { SettingsContext } from "@/providers/settings-provider";
+import { AccordionTrigger } from "./accordion-trigger";
 
 interface AccordionBlockProps extends IFaq {
   idx: number;
@@ -18,45 +13,11 @@ export const AccordionBlock: FC<AccordionBlockProps> = ({
   answer,
   idx,
 }) => {
-  const t = useTranslations();
-  const { iconSize } = useContext(SettingsContext);
-  const [open, setOpen] = useState(false);
-
   return (
     <div title={question}>
-      <button
-        type="button"
-        id={`accordion-${idx}`}
-        className="group inline-flex w-full flex-row items-center justify-between gap-md hyphens-auto py-lg text-start"
-        onClick={() => setOpen(!open)}
-        aria-label={
-          open
-            ? `${t("common.close")} FAQ ${question}`
-            : `${t("common.open")} FAQ ${question}`
-        }
-        aria-expanded={open}
-        aria-controls={`section-${idx}`}
-      >
-        <span className="text-lg underline-offset-4 group-hover:underline">
-          {question}
-        </span>
-        <span className="flex-shrink-0 text-green-600">
-          {open ? (
-            <CrossIcon width={iconSize * 1.5} height={iconSize * 1.5} />
-          ) : (
-            <PlusIcon width={iconSize * 1.5} height={iconSize * 1.5} />
-          )}
-        </span>
-      </button>
-      {open && (
-        <section
-          id={`section-${idx}`}
-          aria-labelledby={`accordion-${idx}`}
-          className="space-y-md pb-lg"
-        >
-          {answer.markdown && <HtmlParser text={answer.markdown} />}
-        </section>
-      )}
+      <AccordionTrigger question={question} idx={idx}>
+        {answer.markdown && <HtmlParser text={answer.markdown} />}
+      </AccordionTrigger>
     </div>
   );
 };
