+++ graphql/index.ts
@@ -1,2 +1,1 @@
-export * from "./containerQuery";
 export * from "./fetcher";
